moved mock cocoon to own file, some added comments
